JS logging formatted; upload test slide.
