+++ javascript/dedicated_server.js
@@ -1,8 +1,7 @@
 // Require modules
-const {errorHandler} = require("./errorHandler");
+const {FILEPATH, errorHandler, getLogger} = require("./helpers");
 
 const express = require('express');
-const winston = require('winston');
 const http = require('http');
 const path = require('path');
 const fs = require('fs');
@@ -13,16 +12,9 @@ const PORT = process.env.PORT || 5000;
 // Definition of constants
 const STUDENT = 1;
 const TEACHER = 2;
-// const FILEPATH = '/mnt/fileserver';
-const FILEPATH = 'D:\\mnt\\fileserver'
 const stuedntsFilename = path.join(FILEPATH, 'registeredInfo', 'registeredStudents.json');
 const trialsFilename = path.join(FILEPATH, 'registeredInfo', 'registeredTrials.json');
-const logger = winston.createLogger({
-    transports: [
-        new winston.transports.Console(),
-        new winston.transports.File({ filename: getLogFilename('dedicated') })
-    ]
-});
+const logger = getLogger('dedicated');
 
 // Run the application
 const app = express();
@@ -393,37 +385,6 @@ function informationPost(req, res) {
         if (err) throw err;
         logger.info('The trials has been saved to file!');
     })
-}
-
-function getLogFilename(servername) {
-    const dedicated = servername.toLowerCase().indexOf('d') >= 0;
-
-    const today = new Date();
-    const logpath = path.join(FILEPATH, 'logs', `${today.getFullYear()}-${today.getMonth() + 1 < 10 ? '0' + (today.getMonth() + 1) : today.getMonth() + 1}-${today.getDate() < 10 ? '0' + today.getDate() : today.getDate()}`);
-    let count = 0;
-
-    if (!fs.existsSync(logpath)) {
-        fs.mkdir(logpath,
-            {recursive: true},
-            (err) => {
-                if (err) throw err;
-            });
-    } else {
-        fs.readdirSync(logpath).forEach(file => {
-            // is js log file?
-            if (file.endsWith('log') && file.toLowerCase().indexOf('js') >= 0) {
-                if (dedicated) {
-                    // filename contains d from dedicated
-                    if (file.toLowerCase().indexOf('d') >= 0) ++count;
-                } else {
-                    // filename does not contain d
-                    if (file.toLowerCase().indexOf('d') < 0) ++count;
-                }
-            }
-        });
-    }
-
-    return path.join(logpath, `${dedicated ? 'dedicated-' : ''}js-${count}.log`);
 }
 
 // ===================================
